refactor: split OAuth pool auth handlers
